move logcardlistview to directory where its used
